chore: add settler interface & refactor
