+++ lcn/settlement.go
@@ -1,0 +1,15 @@
+package lcn
+
+import "github.com/LastL2/cuberollkit/lcn/types"
+
+// Settler is the interface that LcnClient must implement
+type Settler interface {
+	OnStart() error
+	OnStop() error
+	// SubmitBatch submits a batch to the LCN;
+	// Modeled after SubmitBlocks in pkg da
+	SubmitBatch() types.ResultSubmitBatch
+	// RetrieveBatch retrieves a batch from the LCN;
+	// Modeled after RetrieveBlocks in pkg da
+	RetrieveBatch() types.ResultRetrieveBatch
+}
